Prefer inter-shard task list for DROP FKEY commands
This was a bug from utility command pr of the task.
We need to use inter-shard ddl task list function to create task list
for DROP foreign key commands that are run for foreign keys from
reference tables to citus local tables as we do for such ADD foreign
key commands.
This is because, worker placements of reference tables wouldn't have
shard level foreign keys.