Record parallel relation accesses dynamically in the executor
Before this commit, the parellel relation access was based on
the number of tasks, not the actual connections used
by the executor.

With this commit, the executor reacts immediatetly as soon as
the executor requires the second connection to a given node.

The previous logic was problematic for several reasons:

- It had various hacks to make the logic work
- It was not intiutive, easy to follow. It as also fragile, as
  new developments should always consider parallel records
  as opposed to it being automatically
- With the local executor, we were deciding the parallel accesses
  even before we consider local tasks
- The old logic was relying on number of the tasks, which doesn't
  mean much. For example, create_reference_table() creates
  multiple tasks per node, all of which use the same connection.

The downside of this new logic is that the same tx block may work on
some executions and may not for others. The reason is that if the
execution should be prohibited because of foreign keys to reference
tables, the new logic would successfully finish if the execution
requires a single connection only. Still, the fix is straight-forward,
the user just should set `citus.multi_shard_modify_mode` to `sequential`
as instructed by the error hint.