Decide citus table type via catalog if citus has not been loaded
When executing ALTER EXTENSION citus UPDATE commands, IsCitusTable
was always returning false for any existing Citus table as we cannot
access
or build cache entries at that time.

For that purpose, if CitusHasBeenLoaded() returns false and if we
couldn't find
or create a cache entry for given relation, then we should scan
pg_dist_partition
as well instead of directly returning false.