Add component governance config
This config is used to generate components on ADO(Azure devops).
Currently this might not be super useful because we don't really use ADO
but when run, we can see the warnings/issues with components that we
use. A component is like a dependency basically.

(cherry picked from commit 7c11aa124b2123e141121e245d08c62849df58b1)