Detect vars referencing outer queries in local-dist table
We were checking just range table list to detect local-dist table joins
but there can be vars referencing to outer queries which belong to a
distributed table.

The current logic is we keep track of rtable list when we are descending
in the query tree. At any level, we find the range table entries that
are references by the vars of the current subquery. Then we use this to
determine if we should convert local-distributed tables to a subquery.