Consider coordinator in intermediate result optimization
It seems that we were not considering the case where coordinator was
added to the cluster as a worker in the optimization of intermediate
results.

This could lead to errors when coordinator was added as a worker.

(cherry picked from commit 24e60b44a1d0ea95859ad7379b7a6557862f173e)