Fix a bug where one could create a foreign key between non-colocated tables
Due to the scoping and only setting the variable to true when matchend, not
false when not matched, it was possible for a foreign key to be created on
non-colocated tables.

This patch changes the scope and uses an always assign the test result to
the variable we prevent users from creating foreign keys between non-
colocated tables.